iterations/3-mcp-tool/requirements.txt
gradio
haystack-ai
python-dotenv
"mcp[cli]"
httpx
